Organizando

+++ 2021.05.26 Lista de Exercícios 01 - Estruturas de Controle e de Decisão/03. Estrutura Sequencial.js
@@ -3,6 +3,6 @@
 
 //Apresentando variáveis:
 input = require("prompt-sync")();
-let farenheit = parseInt(input("Insira uma temperatura em farenheit: "));
+let farenheit = parseFloat(input("Insira uma temperatura em farenheit: "));
 let celsius = (farenheit - 32) / 1.8;
 console.log(farenheit, 'ºF', '=', celsius.toPrecision(3), 'ºC');
